Refactor credential issue components
