+++ frontend/src/components/helper/CredentialTypeTabs.vue
